MockPrep.ai/ai/requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
gunicorn==21.2.0
flask-cors

SQLAlchemy==2.0.49
psycopg2-binary==2.9.11
alembic==1.13.1
redis==5.0.1

groq
langchain
langchain_groq
langchain-huggingface

python-dotenv==1.0.1
marshmallow==3.21.1
requests==2.31.0
pdfplumber
python-docx

numpy
scipy
pocket-tts

# Face analysis & image extraction
deepface
opencv-python
opencv-python-headless
PyMuPDF
tensorflow
tf-keras
mtcnn
retina-face

# Observability
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
prometheus-client

# Testing
pytest
pytest-cov
pytest-mock
